+++ lib/cocooned/railtie.rb
@@ -7,7 +7,12 @@ module Cocooned
   class Railtie < ::Rails::Engine # :nodoc:
     initializer 'cocooned.initialize' do |_app|
       ActiveSupport.on_load :action_view do
-        ActionView::Base.include Cocooned::Helpers::Tags, Cocooned::Helpers::Containers
+        ActionView::Base.include Cocooned::Helpers::Tags,
+                                 Cocooned::Helpers::Tags::Add,
+                                 Cocooned::Helpers::Tags::Down,
+                                 Cocooned::Helpers::Tags::Remove,
+                                 Cocooned::Helpers::Tags::Up,
+                                 Cocooned::Helpers::Containers
       end
     end
   end
